Fixes for the Q and data vector tests
The tests for the data and Q vectors have been
included. This caused several errors.

For the Q vector most of the fixes are connected
with the highest harmonic number supported that
was not able to compile with the previous structure.

For data vector, the get phi was not there and still
is not there the get id.

Still problems keeping track of the number of components
in the Q vector.