Process, build Q and calibrate incorporated.
The process corrections has been incorporated to
the track and channel detector configurations. It has
been made pure virtual for the base class to separate
both processes. WE NEED TO FOLLOW UP THIS.

Build Q vector has been incorporated into the detector
configurations. Pure virtual in the base class.

The storage of the calibration method has been incorporated
into the base class of detector configurations. It is used
at the end of building the Q vector for proper calibration.

Raw Q vector is incorporated into the channel detector
configuration to get the Q vector before input corrections.

Fix for the way the return value is built when transferring
orders.