Preparation of input corrections.
The states are defined as part of the correction
step base.

The reference of the detection configuration is
stored in the correction step but its setting
is only open to friend class detector configuration.

The process flow incorporates the variable container.
It is included by the framework manager.

The detector configuration opens its data vector bank
and event class variables via gets only available to
correction steps as friend class.

Build raw Qn vector is incorporated and the proper handling of
process corrections on channelized detector configuration
is implemented.

Variable container, channel mask and channel group are protected
as const in all histogram methods.

Number of channels, channels mask and channel groups are open
in the channel detector configuration.

Get channel Id is incorporated in data vector. Minor lex fix.