Adaptations for proper recentering work.
Qn Vector.
Introduced the Qn vector name, the quality flag and the Print function.
Fixes for the or / and precedence issues. Fixes for the normalization
routine.

Detector.
Adaptation for the Qn vector name. Introduced the names of the used Qn
vectors. Incorporated a permanent temporary Qn vector builder for
improving the efficiency. Proper use of the Qn quality service. Fix for
the loop ROOT issue. Fix in providing the list of Qn vectors as
output.

Recentering.
Adaptation for the Qn vector name. Introduced the names of the used Qn
vectors. Do thing only when the Qn vector quality is good and if things
are not done mark the Qn vector quality as not good.