Correction step running conditions
Now the correction steps are informed when the framework
is ready with all needed input attached so they could check
if their conditions for collecting data and for being applied
are fulfilled.

This support is being used by twist and rescale with three
subdetectors method to put conditions over the tracks detectors
being used.